requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
django-resized==1.0.3
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
